pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symbolic-ged-library-generator"
version = "0.1.0"
description = "Symbolic PDE dataset generation + tokenization + VAE training/evaluation utilities."
readme = "src/README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
]

[project.optional-dependencies]
ml = [
  "torch",
  "pytorch-lightning",
  "tqdm",
]
analysis = [
  "scikit-learn",
  "matplotlib",
]
dev = [
  "pytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

